// entidad-web/entidad-web/src/app/@data/model/bases/declaracionjurada.ts
export interface Declaracionjurada {
  configuracionId: number;
  descripcion: string;
  descripcionCorta: string;
  entidadId: number;
  estado: string;
  estadoConfiguracion: string;
  estadoRegistro: string;
  fechaCreacion: any;
  fechaModificacion: any;
  maeCabeceraId: number;
  maeDetalleEntidadId: number;
  orden: any;
  referencia: string;
  sigla: string;
  usuarioCreacion: any;
  usuarioModificacion: any;
  idBase: any;
}